Add authentication to the mutations

+++ instructor-code/src/resolvers/Mutation.js
@@ -1,5 +1,6 @@
 import bcrypt from 'bcryptjs';
 import jwt from 'jsonwebtoken';
+import getUserId from '../utils/getUserId';
 require('dotenv').config();
 
 const secret = process.env.JWT_SECRET;
@@ -23,28 +24,31 @@ const Mutation = {
       token: jwt.sign({ userId: user.id }, secret),
     };
   },
-  async deleteUser(parent, args, { prisma }, info) {
+  async deleteUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
     return prisma.mutation.deleteUser(
       {
         where: {
-          id: args.id,
+          id: userId,
         },
       },
       info
     );
   },
-  async updateUser(parent, args, { prisma }, info) {
+  async updateUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
     return prisma.mutation.updateUser(
       {
         where: {
-          id: args.id,
+          id: userId,
         },
         data: args.data,
       },
       info
     );
   },
-  createPost(parent, args, { prisma }, info) {
+  createPost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
     return prisma.mutation.createPost(
       {
         data: {
@@ -53,7 +57,7 @@ const Mutation = {
           published: args.data.published,
           author: {
             connect: {
-              id: args.data.author,
+              id: userId,
             },
           },
         },
@@ -61,7 +65,17 @@ const Mutation = {
       info
     );
   },
-  deletePost(parent, args, { prisma }, info) {
+  async deletePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId,
+      },
+    });
+    if (!postExists) {
+      throw new Error('Unable to delete post');
+    }
     return prisma.mutation.deletePost(
       {
         where: {
@@ -122,6 +136,24 @@ const Mutation = {
       },
       info
     );
+  },
+  async login(parent, args, { prisma }, info) {
+    const user = await prisma.query.user({
+      where: {
+        email: args.data.email,
+      },
+    });
+    if (!user) {
+      throw new Error('Invalid credentials');
+    }
+    const isMatch = await bcrypt.compare(args.data.password, user.password);
+    if (!isMatch) {
+      throw new Error('Unable to login');
+    }
+    return {
+      user,
+      token: jwt.sign({ userId: user.id }, secret),
+    };
   },
 };
